refactor graient
