+++ Project.toml
@@ -22,6 +22,7 @@ Spec = "b8ccf107-3a88-5e0f-823b-b838c6a0f327"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TransformVariables = "84d833dd-6860-57f9-a1a7-6da5db126cff"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [extras]
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
